Update README.md
COVID-19
COVID-19 is the disease caused by SARS-CoV-2, the coronavirus that emerged in December 2019.
COVID-19 can be severe, and has caused millions of deaths around the world as well as lasting health problems in some who have survived the illness.
The coronavirus can be spread from person to person. It is diagnosed with a test.
The best way to protect yourself is to get vaccinated and boosted when you are eligible, follow testing guidelines, wear a mask, wash your hands and practice physical distancing.